+++ docs/changelog.md
@@ -18,6 +18,7 @@ See the [GitHub "Features" project](https://github.com/users/peter88213/projects
 - Improved and extended the age display.
 - Added a duration display to the section properties.
 - Changed the API of PyCalendar.age().
+- Fixed a bug where the word target and word count start value cannot be reset properly.
 
 
 ### Version 5.23.5